renames and moved file filters to package
